refactor(NextApp): merge AppLayout into _app file
