+++ src/components/AppLayout/index.js
@@ -1,16 +1,3 @@
-export { default } from './AppLayout'
-
-
-
-
-
 export { default as withStripe } from './withStripe'
-export { default as getWrappedPage } from './PageWrapper'
 export { default as asPage } from './Page'
 export { default as asAuthenticatedPage } from './AuthenticatedPage'
-
-
-
-
-
-export * from './AppLayout'
